refactored s3 initialisation into separate util file
