+++ utils/s3.js
@@ -1,0 +1,27 @@
+const multer = require("multer");
+const aws = require("aws-sdk");
+const multerS3 = require("multer-s3");
+const dotenv = require("dotenv");
+
+dotenv.config();
+
+const s3 = new aws.S3({
+  accessKeyId: process.env.ACCESS_KEY_ID,
+  secretAccessKey: process.env.SECRET_ACCESS_KEY,
+});
+
+const multerUpload = multer({
+  storage: multerS3({
+    s3,
+    bucket: process.env.BUCKET_NAME,
+    acl: "public-read",
+    metadata: (request, file, callback) => {
+      callback(null, { fieldName: file.fieldname });
+    },
+    key: (request, file, callback) => {
+      callback(null, Date.now().toString());
+    },
+  }),
+});
+
+module.exports = multerUpload;
